Chart styling via max/min, axis labels and adding blank values for in between dates
